# set RFC4122 fields in Hex128