It turns out that there is real life code that cares about the
RFC4122 variant and version bits, who knew?

So set those bits in the Hex128 result, swapping a couple
of bytes first to avoid duplicating UUIDs after 3.5e16 calls to Next.

The effect on performance turns out to be pretty small:

	name      old time/op  new time/op  delta
	Hex128-4  73.3ns ± 1%  75.2ns ± 1%  +2.48%  (p=0.008 n=5+5)